third_party/WebKit/LayoutTests/fast/dom/NamedNodeMap-missing-arguments.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
[
    "getNamedItem()",
    "getNamedItemNS()",
    "getNamedItemNS('http://www.w3.org/2000/svg')",
    "item()",
    "removeNamedItem()",
    "removeNamedItemNS()",
    "removeNamedItemNS('http://www.w3.org/2000/svg')",
    "setNamedItem()",
    "setNamedItemNS()",
].forEach(function(expr)
{
    window.attributes = document.body.attributes;
    shouldThrow("attributes." + expr);
});
</script>
</body>
</html>
